refactor(guild): rename DeleteInvite to RevokeInvite to match the soft-revoke behavior
