--- services/guild/src/Guild.Application/Features/Invites/DeleteInvite/DeleteInviteCommand.cs
@@ -1,6 +1,0 @@
-using Guild.Application.Abstractions.Messaging;
-using Guild.Domain.Results;
-
-namespace Guild.Application.Features.Invites.DeleteInvite;
-
-public sealed record DeleteInviteCommand(long GuildId, string Code) : ICommand<Result>;
